refactor(WebEditor): move the editor canvas to its own /edit route
Root used to double as both the Play tab (nothing loaded) and the
editor canvas (isLoaded), which meant "should root show the editor"
was decided by a store that deliberately never resets on navigation —
correct for a plain refresh (resume your session), wrong for
explicitly clicking Play or a game detail page's "Back to Play" while
a game happened to still be loaded, which kept landing back in the
editor. A ?view=play query-param override fixed that, but then the
layout's separately-computed tab-bar-visibility check (unaware of the
override) went out of sync with what the page actually rendered,
hiding the tab bar while the Play catalog showed underneath — the
"clunky" seam this commit removes at the source instead of patching
again.

- New src/routes/edit/+page.svelte: the editor canvas, moved verbatim
  out of the old root +page.svelte (Toolbar/TreePanel/PropertyEditor/
  modals, the ?game= server-load path, autosave effects).
- Root +page.svelte: no isLoaded branching at all now. PUBLIC_SHOW_HOME
  true -> always <PlayCatalog />. False (textadventures.co.uk) -> pure
  redirect (/edit if loaded, else /open), matching previous behaviour.
- open/+page.svelte: all eight goto(base || "/") post-load calls, plus
  the ?game= server-created-game redirect, now target /edit.
- +layout.svelte: showTabs is a plain pathname check (hidden only on
  /edit) — no more isLoaded dependency, so the layout and the page it
  wraps can no longer disagree about what's showing.
- HomeTabs' Play link and /play/[id]'s "Back to Play" are back to
  plain {base}/ links — ?view=play no longer needed anywhere.

Verified via dev.sh + Playwright, reproducing the reported flow: create
a game (-> /edit) -> browser back (-> /open, tabs visible) -> click
Play tab (-> /, tabs + catalog) -> click a game -> "Back to Play" ->
tabs + catalog, clean URL, no editor. --api-proxy mode (simulating
textadventures.co.uk) still redirects root straight to /open with no
tab bar, unchanged.
