+++ docs/deployment-domains.md
@@ -63,9 +63,9 @@ Two follow-up fixes found by exercising the real deployed catalog against the Ho
 
 **Play/Create homepage at root** — first shipped 2026-07-16 as a standalone static site in `src/Home/`; **merged into WebEditor 2026-07-16** and `src/Home/` deleted. The standalone version worked (catalog rendering, tab switching, the two catalog bugs it surfaced were fixed server-side — see above) but three real gaps showed up fast: the Create tab was a cross-origin link out to a *different app* rather than feeling like "you're already in the editor"; the Play tab had no way to open a local file/URL (that lived only on WasmPlayer's own start screen at `/player/`); and adding a game-detail view (click a card → description/rating/tags/textadventures.co.uk link, with an explicit Play button, rather than launching straight into WasmPlayer) needed real routing/state, which a hand-rolled static page doesn't have. Rather than build a second bespoke app next to WebEditor's existing SvelteKit SPA, Home's Play/Create UI now lives inside it as more routes:
 
-- **Root (`/`)** — Play tab (game catalog) when nothing's loaded, the editor canvas once a game is (same `isLoaded` check as before, just a different "nothing loaded" screen). Gated by a new `PUBLIC_SHOW_HOME` build-time flag (`src/WebEditor/.env`, default `false`) — `true` for play.questviva.com and Electron, unset for textadventures.co.uk's `/questviva` build, which keeps its previous editor-only root behavior (that origin has its own site navigation, not a game-browser homepage).
-- **`/open`** — unchanged (Create tab's content); the Play/Create tab bar (`HomeTabs.svelte`) lives in the root layout, visible on `/` and `/open` (and `/play/[id]`) whenever `PUBLIC_SHOW_HOME` is true and nothing's loaded, hidden once a game is.
-- **`/play/[id]`** — new game-detail route, fetching `api/GameDetails/{id}` (added specifically for this — see below).
+- **Root (`/`)** — always the Play tab (game catalog) when `PUBLIC_SHOW_HOME` is true, full stop. The editor canvas originally lived at root too (shown whenever `isLoaded`), but that broke as soon as Play-tab navigation existed alongside it: `isLoaded` deliberately never resets just from navigating elsewhere (a refresh of `/` should resume your session), so clicking "Play" or a game detail page's "Back to Play" while something was still loaded in the background kept landing back in the editor instead. A `?view=play` query-param override was tried first and fixed that, but then the *layout's* separately-computed tab-bar-visibility check (which didn't know about the override) went out of sync with what the page actually rendered, hiding the tab bar while the Play catalog showed underneath. **Resolved by giving the editor its own route, `/edit`** (moved verbatim out of the old root `+page.svelte`) — root now has no `isLoaded` branching at all, and every place that used to `goto(base || "/")` after loading a game (eight call sites in `open/+page.svelte`, plus the `?game=` server-load entry point) now targets `/edit` instead. `PUBLIC_SHOW_HOME=false` (textadventures.co.uk) unaffected: root there is purely a redirect (`/edit` if loaded, else `/open`), matching its previous behaviour exactly.
+- **`/open`** — unchanged (Create tab's content), except its post-load navigation now targets `/edit`. The Play/Create tab bar (`HomeTabs.svelte`) lives in the root layout, visible on every route except `/edit` — a plain pathname check now, no `isLoaded` involved.
+- **`/play/[id]`** — new game-detail route, fetching `api/GameDetails/{id}` (added specifically for this — see below). Its "Back to Play" link is a plain `{base}/`, same as the tab bar's Play link — no query param needed now that root unconditionally means Play.
 - Catalog/detail fetches live in `$lib/home-catalog.ts`; `source`/`version`/`platform` client info uses `PUBLIC_WEBEDITOR_VERSION` (already baked in at build time — no more runtime `fetch('VERSION')` workaround) and the same `isElectron()` check `runtime.ts` already had. `platform` is declared optional on `window.electronApp` (`electron-types.d.ts`) since `ElectronApp`'s `preload.ts` doesn't populate it yet — reads through automatically once it does.
 - `MAX_ASL_VERSION = 580` in `home-catalog.ts` still needs bumping by hand alongside any Engine change to `GameLoader.cs`'s `Versions` dictionary — same caveat as the static version had, just moved.
 - Local dev: `dev.sh` now starts only WebEditor (port 5174, already proxying `/player` to WasmPlayer via `vite.config.ts`) and WasmPlayer — no separate Home dev server, no `/api`/`/game-resource` proxy workarounds (those existed only to route around Home's dev server not knowing about WasmPlayer's own dev-only quest-config.js override). `--api-proxy` (simulating textadventures.co.uk) also now sets `PUBLIC_SHOW_HOME=false` to match.
